# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AgentTODO API
  description: >
    <!-- @purpose -->AI 驱动的本地优先任务系统 RESTful API，支持可信网络免认证或可选 API Token，无缝对接 AI 助手和命令行技能系统。<!-- /purpose -->
  version: 1.0.0
  contact:
    name: AgentTODO
  license:
    name: MIT
    url: https://github.com/alanhsun/AgentTODO/blob/main/LICENSE
servers:
  - url: http://localhost:3300
    description: 本地开发

# API_TOKEN 未设置时允许匿名访问；设置后使用以下任一认证方式。
security:
  - bearerToken: []
  - basicAuth: []
  - {}
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
    basicAuth:
      type: http
      scheme: basic
  schemas:
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        task_count:
          type: integer
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        due_date:
          type: string
          format: date
          nullable: true
        recurrence:
          type: string
          enum: [none, daily, weekly, monthly]
        recurrence_end:
          type: string
          format: date
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
...
    TaskCreateInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
          default: todo
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        due_date:
          type: string
          format: date
          nullable: true
        recurrence:
          type: string
          enum: [none, daily, weekly, monthly]
          default: none
        recurrence_end:
          type: string
          format: date
          nullable: true
        tags:
          type: array
          items:
            type: integer
          uniqueItems: true
        subtasks:
          type: array
          items:
            oneOf:
              - type: string
              - type: object
                required: [title]
                properties:
                  title: { type: string, maxLength: 255 }
    TaskUpdateInput:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        due_date:
          type: string
          format: date
          nullable: true
        recurrence:
          type: string
          enum: [none, daily, weekly, monthly]
        recurrence_end:
          type: string
          format: date
          nullable: true
        tags:
          type: array
          items:
            type: integer
          uniqueItems: true
    BatchInput:
      type: object
      required: [action, ids]
      properties:
        action:
          type: string
          enum: [update_status, update_priority, delete]
        ids:
          type: array
          items:
            type: integer
          minItems: 1
        value:
          type: string
          description: 新的状态或优先级值（action为delete时不需要）
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string
paths:
...
  /api/tasks:
    get:
      summary: 获取任务列表
      tags: [Tasks]
      parameters:
        - name: status
          in: query
          schema:
            type: string
          description: "按状态筛选，多个用逗号分隔 (todo,in_progress,done)"
        - name: priority
          in: query
          schema:
            type: string
          description: "按优先级筛选，多个用逗号分隔"
        - name: tag
          in: query
          schema:
            type: string
          description: "按标签ID筛选，多个用逗号分隔"
        - name: search
          in: query
          schema:
            type: string
          description: 搜索标题和描述
        - name: due_before
          in: query
          schema:
            type: string
            format: date
          description: 截止日期早于
        - name: due_after
          in: query
          schema:
            type: string
            format: date
          description: 截止日期晚于
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, due_date, priority, title, status]
            default: created_at
          description: 排序字段
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: 排序方向
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: 页码
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: 每页条数
      responses:
        '200':
          description: 任务列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: 创建任务
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateInput'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
  /api/tasks/{id}:
    get:
      summary: 获取单个任务
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: 任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: 更新任务
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateInput'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: 任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: 删除任务
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: 任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tasks/batch:
    post:
      summary: 批量操作
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchInput'
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  affected:
                    type: integer
        '400':
          description: 参数错误
...